Fixed some logical errors when saving and deleting scenarios. #14
